requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# Новые зависимости для Step 1
celery==5.3.4
minio==7.2.0
pika==1.3.2
redis==5.0.1
flower==2.0.1

# Step 2: Аутентификация и авторизация  
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.11.7
